+++ src/App.js
@@ -1,23 +1,13 @@
-import logo from './logo.svg';
 import './App.css';
+import './fonts.css';
+import "@patternfly/react-core/dist/styles/base.css";
+
+import VersionDf from './components/VersionDf';
 
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      <VersionDf />
     </div>
   );
 }
